更新了一些错误检测

+++ src/main/java/org/laorui_out/habit_former/poster/service/DeletePosterService.java
@@ -1,16 +1,27 @@
 package org.laorui_out.habit_former.poster.service;
 
 import jakarta.annotation.Resource;
+import org.laorui_out.habit_former.bean.PosterBean;
 import org.laorui_out.habit_former.mapper.PosterMapper;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class DeletePosterService {
     @Resource
     PosterMapper posterMapper;
+    @Resource
+    PosterService posterService;
 
     //根据帖子ID删除帖子
     public Boolean deletePoster(int posterID){
+
+        //对于输入posterID是否存在的判断
+        PosterBean posterTestBean = posterMapper.getPosterById(posterID);
+        if(posterTestBean == null){
+            return false;
+        }
         try {
             // 删除 posterpicture 表中的记录
             boolean pictureDeleted = posterMapper.deletePosterPictureByPosterId(posterID);
